feat : git new project Ready (sprint mission 5)
