+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = '5-spring-mission'
+rootProject.name = 'discodeit'
