+++ src/main/java/com/sprint/mission/discodeit/DiscodeitApplication.java
@@ -1,0 +1,13 @@
+package com.sprint.mission.discodeit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DiscodeitApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DiscodeitApplication.class, args);
+	}
+
+}
