+++ src/main/java/com/sprint/mission/discodeit/JavaApplication.java
@@ -1,0 +1,19 @@
+package com.sprint.mission.discodeit;
+
+import com.sprint.mission.discodeit.mainview.MainViewBasic;
+import com.sprint.mission.discodeit.mainview.MainViewFile;
+import com.sprint.mission.discodeit.mainview.MainViewJCF;
+
+public class JavaApplication {
+
+
+	public static void main(String[] args) {
+		MainViewJCF mainViewJCF = new MainViewJCF();
+		MainViewFile mainViewFile = new MainViewFile();
+		MainViewBasic mainViewBasic = new MainViewBasic();
+		mainViewJCF.mainMenu();		//JCF 테스트 용
+		// mainViewFile.mainMenu(); 	//File 테스트 용
+		// mainViewBasic.mainMenuBasic();
+
+	}
+}
